Section sync between editors
