+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "simple-text-editor-app",
-  "version": "0.5.0",
+  "version": "0.6.0",
   "homepage": "https://simple-text-editor-app.netlify.app/",
   "private": false,
   "dependencies": {
@@ -11,7 +11,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-scripts": "4.0.3",
-    "simple-text-editor-rcl": "2.1.0",
+    "simple-text-editor-rcl": "2.2.0",
     "use-file-picker": "^1.4.0",
     "web-vitals": "^1.0.1"
   },
