+++ src/App.js
@@ -1,22 +1,10 @@
 
-import TextFileEditor from './components/TextFileEditor';
+import Layout from './components/Layout';
 
 function App() {
-  const workspaceStyle = {
-    width: '100%',
-    display: 'flex',
-    alignItems: 'flex-start',
-  };
   return (
     <div className="App">
-      <div style={workspaceStyle}>
-        <div style={{ width: '100%', overflowY: 'scroll', padding: '0 1em', }}>
-          <TextFileEditor editable={false} />
-        </div>
-        <div style={{ width: '100%', overflowY: 'scroll', padding: '0 1em', }}>
-          <TextFileEditor />
-        </div>
-      </div>
+      <Layout />
     </div>
   );
 }
